i9300: audio: bring back removed microphone fix
http://review.cyanogenmod.org/61104 has reverted http://review.cyanogenmod.org/51085
which has bought back issues with microphone after a voice call. After a voice call,
the microphone is either not working or the recorded audio is very faint and not
audible.

The Input device has to be explicitly set to AUDIO_DEVICE_NONE after a voice call,
since force_all_standby() wouldn't work for a voice call usecase.

More details about the fix is included in the gerrit comments for the actual patch.

Change-Id: I87c63b0ec4390ec3e5703e1de5804c10289ee423

Conflicts:
	audio/audio_hw.c